Added different gameboards
Both Menu buttons go to different gameboards set to different modes. Alot of code is repeated so could be made redendant in future by creating the board the same and than just running different listener functions depending on gamemode.

Also added an array for the players go to go into to allow the algorithm to see the current gameboard. Seems to be having problems with resetting it upon starting a new game+ not working for all buttons. Therefore the array is still a workin progress, to test the array play 2 player gamemode with the array to fill it and print the array to test it could improve the printing statemenmts to imporve visual feedback of what is happening