Definitly didnt forget about draws
So, did definitely remember to put in a case in the if statement if there was a draw. Like it was there the whole time, dont worry about it.
Also fixed the gameboard array so now it will start blank and fill the array with data when a player performs a go.
Buttons can be disabled after X's go again and the minimax algorithm can be made.