IVE DONE IT!!!!!!
Ive actually done it, (i think).

The AI is blocking my moves not which is great, the ai is not making stupid decisions either. But bc i cba to test it thoroughly, ill just send to someone and tell them to try and win.

But tbh just need to add an ez of life feature to switch who goes first and hopefully it does not break it.